refactored guitar component split guitar component into many subcomponents split guitarForScrollModel into many subcomponents
